노트북 성능이 낮으니 할 맛이 안나네
