+++ db.json
@@ -18,132 +18,20 @@
       "createdAt": "Sun Dec 03 2023 21:50:31 GMT+0900 (한국 표준시)"
     },
     {
-      "id": "7ea8795b-1909-4eb9-a8b8-cc7bf898aa3b",
+      "id": "5e04c128-32ad-45bd-b118-cd014a841c7b",
       "nickname": "qweqwe",
-      "content": "qweeedddd",
+      "content": "ddddd",
       "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
       "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 00:26:49 GMT+0900 (한국 표준시)"
+      "createdAt": "Mon Dec 04 2023 11:44:30 GMT+0900 (한국 표준시)"
     },
     {
-      "id": "4e16a478-4146-4650-b269-fcb821cdc726",
+      "id": "8f6c9159-8252-433c-ac49-575d3d6ac866",
       "nickname": "qweqwe",
-      "content": "wwrerttfff",
+      "content": "ddddd",
       "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
       "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 00:27:24 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "ed64b92c-bd4d-4ce2-8bd0-78657d8bc6a0",
-      "nickname": "qweqwe",
-      "content": "qqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 00:31:28 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "25caa5a4-abfc-47f4-949d-294b6771b6ee",
-      "nickname": "qweqwe",
-      "content": "qweqwe",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 00:34:39 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "decdc0d1-231c-440c-8d01-b9eff580a6a0",
-      "nickname": "qweqwe",
-      "content": "qqqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "윈터",
-      "createdAt": "Mon Dec 04 2023 09:12:22 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "6c688aaf-9e94-43bf-8252-4d1ee87494aa",
-      "nickname": "qweqwe",
-      "content": "qqqqqqqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "윈터",
-      "createdAt": "Mon Dec 04 2023 09:13:23 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "706be50a-fe05-44d9-ab0b-7ebcc898bd1e",
-      "nickname": "qweqwe",
-      "content": "qqqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "닝닝",
-      "createdAt": "Mon Dec 04 2023 09:14:24 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "423e7dda-79fc-4f87-b73d-8e8b380aad5e",
-      "nickname": "qweqwe",
-      "content": "qqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 09:15:06 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "dd1e6403-71f5-4533-b21b-6ec7b47d12f5",
-      "nickname": "qweqwe",
-      "content": "qqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 09:16:50 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "98261f9a-a95d-4b1c-a50f-399cc63ed7bf",
-      "nickname": "qweqwe",
-      "content": "qqqqqqqqqqqqqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 09:17:39 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "e29432f2-d4fa-4108-ad78-22980e893da0",
-      "nickname": "qweqwe",
-      "content": "qweqwee",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 09:20:39 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "86a41772-f270-47ce-a95b-4a08358ac2aa",
-      "nickname": "qweqwe",
-      "content": "qqqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "윈터",
-      "createdAt": "Mon Dec 04 2023 09:21:00 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "a0178723-bc55-479b-b854-03a84f952c2b",
-      "nickname": "qweqwe",
-      "content": "wwwwww",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "윈터",
-      "createdAt": "Mon Dec 04 2023 09:21:08 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "52188e40-390c-4724-b0d8-8d9d3a4bdadd",
-      "nickname": "qweqwe",
-      "content": "qqqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 09:24:01 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "d779c5c6-1fc5-4fed-b4a3-9c92852e78cd",
-      "nickname": "qweqwe",
-      "content": "qqqq",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 09:29:05 GMT+0900 (한국 표준시)"
-    },
-    {
-      "id": "f7dc6716-c10f-4d1b-af8e-e6aae6187c8b",
-      "nickname": "qweqwe",
-      "content": "qweqwww",
-      "avatar": "https://images.unsplash.com/photo-1561962534-50ff147395c3?w=125&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8ODZ8fCVFQyU5RCVCNSVFQiVBQSU4NSVFQyU5RCU5OHxlbnwwfHwwfHx8MA%3D%3D",
-      "writedTo": "카리나",
-      "createdAt": "Mon Dec 04 2023 09:29:14 GMT+0900 (한국 표준시)"
+      "createdAt": "Mon Dec 04 2023 11:45:47 GMT+0900 (한국 표준시)"
     }
   ]
 }
